+++ src/app/(home)/page.tsx
@@ -155,12 +155,12 @@ export default function HomePage() {
                   transition={{ delayChildren: 0.1 }}
                 >
                   <motion.div className="flex flex-col items-center w-full" variants={defaultAnimations} transition={{ duration: 2 }}>
-                    <h2 className="mb-12 font-bold tracking-wide sm:text-4xl">View with examples</h2>
+                    <h2 className="mb-12 font-bold tracking-wide sm:text-4xl">View with docs</h2>
                     <Button
                       variant="outline"
                       className="h-auto p-0 text-base font-medium uppercase transition-colors border border-blue-600 rounded-lg hover:bg-inherit hover:border-blue-600/80"
                     >
-                      <Link href="/examples" className="flex h-12 px-12 py-3 [&>svg]:hover:animate-wobble ">
+                      <Link href="/docs" className="flex h-12 px-12 py-3 [&>svg]:hover:animate-wobble ">
                         <span className="mr-3">Learn more</span>
                         <MoveRight />
                       </Link>
