+++ src/app/docs/layout.tsx
@@ -8,17 +8,17 @@ import { ChevronRight } from "lucide-react";
 import GlobalNav from "@/components/ui/global-nav";
 import Breadcrumb from "@/components/ui/breadcrumb";
 
-export default function ExamplesLayout({ children }: { children: React.ReactNode }) {
+export default function Layout({ children }: { children: React.ReactNode }) {
   const pathname = usePathname();
 
-  const isIntro = pathname === "/examples";
+  const isIntro = pathname === "/docs";
 
   useEffect(() => {
     document.body.classList.add('scroll');
   }, [])
 
   return (
-    <div className="flex-1">
+    <>
       <div className="border-b">
         <div className="container flex-1 items-start md:grid md:grid-cols-[220px_minmax(0,1fr)] md:gap-6 lg:grid-cols-[240px_minmax(0,1fr)] lg:gap-10">
           <GlobalNav />
@@ -27,7 +27,7 @@ export default function ExamplesLayout({ children }: { children: React.ReactNode
               {isIntro ? (
                 <div className="-mt-4 mb-7 md:mb-10">
                   <div className="flex flex-wrap items-center gap-2 text-sm">
-                    <span className="overflow-hidden text-ellipsis whitespace-nowrap text-muted-foreground">Examples</span>
+                    <span className="overflow-hidden text-ellipsis whitespace-nowrap text-muted-foreground">docs</span>
                     <span>
                       <ChevronRight className="w-4 h-4" />
                     </span>
@@ -42,6 +42,6 @@ export default function ExamplesLayout({ children }: { children: React.ReactNode
           </main>
         </div>
       </div>
-    </div>
+    </>
   );
 }
