+++ src/app/docs/nextjs/routing/nesting-layouts/[category]/layout.tsx
@@ -1,17 +1,23 @@
-import { categories } from "@/data/categories";
+"use client";
 
-import { BadgeLayout } from "@/components/layout/badge-layout";
-import { AddressBar } from "@/components/ui/address-bar";
-import { TabGroup } from "@/components/ui/tab-group";
+import { movieCategories, apparelCategory } from "@/data/categories";
+
+import { useParams } from "next/navigation";
+
 import { ClickCounter } from "@/components/ui/click-counter";
+import { BadgeLayout } from "@/components/layout/badge-layout";
+import { TabGroup } from "@/components/ui/tab-group";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
+  const params = useParams();
+
+  const categories = params.category === "movie" ? movieCategories : apparelCategory;
+
   return (
-    <BadgeLayout display="Home Layout">
-      <AddressBar />
+    <BadgeLayout display="Children Layout">
       <div className="flex items-center justify-between px-2">
         <TabGroup
-          path="/examples/nextjs/routing/route-groups"
+          path={`/docs/nextjs/routing/nesting-layouts/${params.category}`}
           items={[
             {
               key: "home",
@@ -22,16 +28,6 @@ export default function Layout({ children }: { children: React.ReactNode }) {
               name: c.name,
               slug: c.key,
             })),
-            {
-              key: "order",
-              name: "Order",
-              slug: "order",
-            },
-            {
-              key: "campaign",
-              name: "Campaign",
-              slug: "campaign",
-            },
           ]}
         />
         <ClickCounter />
