+++ src/app/docs/nextjs/routing/nesting-layouts/layout.tsx
@@ -1,17 +1,17 @@
 import { categories } from "@/data/categories";
 
-import { BadgeLayout } from "@/components/layout/badge-layout";
-import { AddressBar } from "@/components/ui/address-bar";
-import { TabGroup } from "@/components/ui/tab-group";
 import { ClickCounter } from "@/components/ui/click-counter";
+import { BadgeLayout } from "@/components/layout/badge-layout";
+import { TabGroup } from "@/components/ui/tab-group";
+import { AddressBar } from "@/components/ui/address-bar";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
-    <BadgeLayout display="Commerce Layout">
+    <BadgeLayout display="Layout">
       <AddressBar />
       <div className="flex items-center justify-between px-2">
         <TabGroup
-          path="/examples/nextjs/routing/route-groups"
+          path="/docs/nextjs/routing/nesting-layouts"
           items={[
             {
               key: "home",
@@ -22,16 +22,6 @@ export default function Layout({ children }: { children: React.ReactNode }) {
               name: c.name,
               slug: c.key,
             })),
-            {
-              key: "order",
-              name: "Order",
-              slug: "order",
-            },
-            {
-              key: "campaign",
-              name: "Campaign",
-              slug: "campaign",
-            },
           ]}
         />
         <ClickCounter />
