+++ src/app/docs/nextjs/routing/route-groups/(commerce)/[category]/page.tsx
