+++ src/app/docs/nextjs/routing/route-groups/(home)/page.tsx
